backend/requirements.txt
aiosmtplib==4.0.2
alembic==1.16.5
boto3==1.40.10
fastapi==0.116.1
fastapi-pagination==0.14.0
GeoAlchemy2==0.18.0
google-cloud-tasks>=2.16.0
greenlet==3.2.4
gunicorn==23.0.0
Jinja2==3.1.6
matplotlib==3.10.6
openai==1.102.0
psycopg[binary]==3.2.10
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
shapely==2.1.1
SQLAlchemy==2.0.43
twilio==9.7.2
uvicorn==0.35.0
